+++ CHANGES.rst
@@ -5,7 +5,8 @@ CHANGES
 5.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Fixed browser to only follow redirects for HTTP statuses
+  301, 302, 303, and 307; not other 30x statuses such as 304.
 
 
 5.1 (2017-01-31)